requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
openai>=1.0.0
pandas>=1.5.0
numpy>=1.21.0
streamlit>=1.24.0
python-dotenv>=0.19.0
streamlit-agraph>=0.0.7
langchain-openai>=0.0.5
langchain-core>=0.1.15